Refactoring invader
